use map[string]interface to unmarshal
We have to use map[string]interface{} to handle the discrepancy between
java json library and golang json library. We have to unmarshal the
result to map[string]interface first to avoid unmarshall error. Then go
throuth golang json unmarshal and marshall cycle to convert it to the
right struct.